+++ src/game/CompactRules/constants/index.ts
@@ -1,3 +1,4 @@
 export * from "./standardGaits";
 export * from "./hexGaits";
+export * from "./nimbusGaits";
 export * from "./tokens";
